[ICU-13177] Removes old (ICU 59) number formatting middleware since it has been wholly replaced by the property mapper function. This change also removes the IProperties polymorphism and makes all remaining classes reference Properties directly. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/40352 Former-commit-id: 207978e248c2ead189aade36eddf4b84b8aa4636

+++ icu4j/main/classes/core/src/newapi/impl/SkeletonBuilder.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 import java.math.MathContext;
 import java.math.RoundingMode;
 
-import com.ibm.icu.impl.number.formatters.PaddingFormat.PadPosition;
+import com.ibm.icu.impl.number.ThingsNeedingNewHome.PadPosition;
 import com.ibm.icu.text.CompactDecimalFormat.CompactStyle;
 import com.ibm.icu.text.DecimalFormatSymbols;
 import com.ibm.icu.text.MeasureFormat.FormatWidth;
@@ -461,9 +461,9 @@ public final class SkeletonBuilder {
       if (c1 == ',') {
         char c2 = safeCharAt(skeleton, offset++);
         char c3 = safeCharAt(skeleton, offset++);
-        result = GroupingImpl.getInstance((short) (c0 - '0'), (short) (c2 - '0'), c3 == '&');
+        result = GroupingImpl.getInstance((byte) (c0 - '0'), (byte) (c2 - '0'), c3 == '&');
       } else {
-        result = GroupingImpl.getInstance((short) (c0 - '0'), (short) (c0 - '0'), c1 == '&');
+        result = GroupingImpl.getInstance((byte) (c0 - '0'), (byte) (c0 - '0'), c1 == '&');
       }
     } else {
       StringBuilder sb = new StringBuilder();
